Initial implementation, extracted from sbt-crossproject.
In sbt-crossproject, `%%%` based its decisions on the value of
`crossPlatform`. Obviously, it cannot do so here, since
`crossPlatform` is a concept of sbt-crossproject. Instead, we have
a dedicated setting `platformDepsCrossVersion`, which directly
holds the `CrossVersion` to use in generated `ModuleID`s.

sbt plugins for platforms, such as sbt-scalajs and
sbt-scala-native, only need to set this setting to an appropriate
`CrossVersion` for `%%%` to work as expected.